modal-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
Pillow>=10.0.0
modal>=1.1.4
aiortc>=1.7.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
